Lint + gas snapshot update

+++ packages/contracts-bedrock/test/RoundRobinLeaderElection.t.sol
@@ -20,7 +20,9 @@ contract RoundRobinLeaderElectionTest is Test {
         RoundRobinLeaderElection leaderImpl = new RoundRobinLeaderElection();
 
         vm.prank(msg.sender);
-        proxy.upgradeToAndCall(address(leaderImpl), abi.encodeCall(leaderImpl.initialize, (owner, N_PARTICIPANTS, N_SLOTS_PER_LEADER)));
+        proxy.upgradeToAndCall(
+            address(leaderImpl), abi.encodeCall(leaderImpl.initialize, (owner, N_PARTICIPANTS, N_SLOTS_PER_LEADER))
+        );
 
         leaderContract = RoundRobinLeaderElection(address(proxy));
 
@@ -30,7 +32,7 @@ contract RoundRobinLeaderElectionTest is Test {
             address addr = vm.addr(i);
             leaderContract.addParticipant(addr);
         }
-        assertEq(N_PARTICIPANTS*N_SLOTS_PER_LEADER,leaderContract.horizon());
+        assertEq(N_PARTICIPANTS * N_SLOTS_PER_LEADER, leaderContract.horizon());
     }
 
     function test_addParticipant_participantsAreAdded_success() external {
@@ -56,12 +58,12 @@ contract RoundRobinLeaderElectionTest is Test {
         }
 
         // When the contract is deployed the leader is the first participant
-        for (uint i=0;i<N_PARTICIPANTS;i++){
-            for (uint j=0;j<N_SLOTS_PER_LEADER;j++){
-                uint blockNumber = DEPLOYMENT_BLOCK_NUMBER + i*N_SLOTS_PER_LEADER + j;
-                assertTrue(leaderContract.isCurrentLeader(vm.addr(i+1), blockNumber));
-                blockNumber = DEPLOYMENT_BLOCK_NUMBER + 3*leaderContract.horizon() + i*N_SLOTS_PER_LEADER + j;
-                assertTrue(leaderContract.isCurrentLeader(vm.addr(i+1), blockNumber));
+        for (uint256 i = 0; i < N_PARTICIPANTS; i++) {
+            for (uint256 j = 0; j < N_SLOTS_PER_LEADER; j++) {
+                uint256 blockNumber = DEPLOYMENT_BLOCK_NUMBER + i * N_SLOTS_PER_LEADER + j;
+                assertTrue(leaderContract.isCurrentLeader(vm.addr(i + 1), blockNumber));
+                blockNumber = DEPLOYMENT_BLOCK_NUMBER + 3 * leaderContract.horizon() + i * N_SLOTS_PER_LEADER + j;
+                assertTrue(leaderContract.isCurrentLeader(vm.addr(i + 1), blockNumber));
             }
         }
     }
@@ -106,7 +108,6 @@ contract RoundRobinLeaderElectionTest is Test {
             abi.encodePacked([true, true, true, false, false, false, false, false, false, false, false, false])
         );
 
-
         // Happy case for third leader
         leader = vm.addr(3);
         blockNumber = DEPLOYMENT_BLOCK_NUMBER + 3;
@@ -118,4 +119,3 @@ contract RoundRobinLeaderElectionTest is Test {
         );
     }
 }
-
